+++ packages/table-render/src/Search.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useTable } from './hooks';
 import { Button } from 'antd';
 import SearchForm from 'form-render';
-import { SearchApi } from './interface';
+import { SearchProps } from './interface';
 
 const SearchBtn = ({
   clearSearch,
@@ -59,27 +59,6 @@ const MySearchBtn = ({
     />
   );
 };
-
-export interface SearchProps<RecordType> {
-  debug?: boolean;
-  searchBtnStyle?: React.CSSProperties;
-  searchBtnClassName?: string;
-  api?: SearchApi<RecordType> | Array<SearchApi<RecordType>>;
-  displayType?: any;
-  propsSchema?: any;
-  className?: string;
-  style?: React.CSSProperties;
-  schema?: any;
-  hidden?: boolean;
-  searchOnMount?: boolean | unknown;
-  searchBtnRender?: (
-    submit: Function,
-    clearSearch: Function
-  ) => React.ReactNode[];
-  onSearch?: (search: any) => any;
-  afterSearch?: (params: any) => any;
-  widgets?: any;
-}
 
 const Search: <RecordType extends object = any>(props: SearchProps<RecordType>) => React.ReactElement = (props) => {
   const { searchBtnRender, searchBtnStyle, searchBtnClassName } = props;
